[WIP] 0.7.0

+++ src/test/scala/fr/acinq/hc/app/HostedUpdatesDbSpec.scala
@@ -1,16 +1,16 @@
 package fr.acinq.hc.app
 
 import fr.acinq.bitcoin.Crypto.PrivateKey
-import fr.acinq.eclair._
-import slick.jdbc.PostgresProfile.api._
-import fr.acinq.eclair.wire.protocol.LightningMessageCodecs._
 import fr.acinq.bitcoin.{Block, ByteVector32, ByteVector64, Crypto}
+import fr.acinq.eclair._
 import fr.acinq.eclair.router.Announcements
+import fr.acinq.eclair.wire.protocol.LightningMessageCodecs._
 import fr.acinq.hc.app.Tools.DuplicateHandler
 import fr.acinq.hc.app.db.{Blocking, HostedUpdatesDb, PreimagesDb, Updates}
 import fr.acinq.hc.app.network.{CollectedGossip, PHC}
 import org.scalatest.funsuite.AnyFunSuite
 import scodec.bits.BitVector
+import slick.jdbc.PostgresProfile.api._
 
 
 class HostedUpdatesDbSpec extends AnyFunSuite {
@@ -48,8 +48,8 @@ class HostedUpdatesDbSpec extends AnyFunSuite {
     assert(channelAnnouncementCodec.decode(BitVector.fromValidHex(res1._3)).require.value == channel1)
     assert(res1._4.isEmpty)
 
-    Blocking.txWrite(Updates.update1st(channel_update_1.shortChannelId.toLong, channelUpdateCodec.encode(channel_update_1).require.toHex, channel_update_1.timestamp), HCTestUtils.config.db)
-    Blocking.txWrite(Updates.update2nd(channel_update_2.shortChannelId.toLong, channelUpdateCodec.encode(channel_update_2).require.toHex, channel_update_1.timestamp), HCTestUtils.config.db)
+    Blocking.txWrite(Updates.update1st(channel_update_1.shortChannelId.toLong, channelUpdateCodec.encode(channel_update_1).require.toHex, channel_update_1.timestamp.toLong), HCTestUtils.config.db)
+    Blocking.txWrite(Updates.update2nd(channel_update_2.shortChannelId.toLong, channelUpdateCodec.encode(channel_update_2).require.toHex, channel_update_1.timestamp.toLong), HCTestUtils.config.db)
 
     val res2 = Blocking.txRead(Updates.model.result, HCTestUtils.config.db).head
     assert(channelUpdateCodec.decode(BitVector.fromValidHex(res2._4.get)).require.value == channel_update_1)
